Implement SNI sniffing based on github.com/inconshreveable/go-vhost
Changes the behavior slightly for the "ConnectAccept" case, which will
always send a "200 OK", but will shut down the connection if the remote end
doesn't answer.

The SniffSNI property on the ProxyHttpServer is a proposition, so we could
handle both cases.  For those who don't need SNI, we could skip the overhead,
but those who enable it, would accept the altered "ConnectAccept".

The code doesn't deal with this switch, but if the idea is accepted, I
could rework it to make it a nicely switchable feature.