Fix panic if CONNECT transport is not a net.TCPConn
handleHttps would panic if Dial returned something other than a
net.TCPConn which was then used for a non MITM CONNECT request.

For example the golang.org/x/crypto/ssh package returns a net.Conn
when opening a tunnel through an ssh connection.  This is backed
by a wrapped ssh.Channel which also implements CloseWrite() so using
a type assertion achieves the same result in a portable fashion.