Modified version of github.com/elazarl/goproxy
Important changes:

Fixes #256:  connections are closed after usage, so FD-count remains
congruent with the actual amount of open connections.

Added ProxyHttpServer.Signer of type
	type Signer func(ca *tls.Certificate, hostname []string) (*tls.Certificate, error)
If ProxyHttpServer.Signer is set, the goproxy uses that function to
retrieve Certificates for TLS-interception.  This allows consumers of
goproxy to implement f.e. caching of certificates.

Minor changes:

- adjusted import-path to github.com/oec/goproxy
- go fmt on all files