Switch CertStorage interface to single function
Rather than having Store() and Load() as separate functions, Fetch() is
provided with the key and a function that returns the value to cache,
and it handles the store and load internally.

A simple storage implementation may mimic the previous behaviour, but
the problem with this is many concurrent requests will result in the
certificate being generated many times in parallel.

An alternative implementation of the certificate storage might lock on
the key when generating the certificate, meaning only one certificate
will be generated and other clients will wait until it exists.