connect/tls-mitm: Read body and reset buffer
After successful handshake in a HTTPS MITM scenario, it is not
sufficient call http.ReadRequest(clientTLS).  The body -or parts of it-
might still be in the bufio.Buffer and not processed yet.  The
subsequent call to http.ReadRequest will than fail because it doesn't
find a proper request-line in the buffer.

This patch reads the complete body from the request and resets the
bufio.Buffer so that the next call to http.ReadRequest() can work with a
clean plate.

[Unfortunately I haven't managed to create an test case for it as the
http.Client seems to open new CONNECTs to access the test-TLS-server.
Still working on it.]