Fix: outgoing req.Host must have no port during intercept
When intercepting a CONNECT, we read the request from the client in a
loop and create another request object.  Before this commit, we had set
the .Host of this new request with the .Host from the original
CONNECT-request, which might contain a port number.  Such a request to
the outgoing server failed with 500.

With this commit, the proxy does not set .Host anymore on the outgoing
request (but still sets .URL.Host).